+++ Public/FeeRTOS_Timer.h
@@ -1,6 +1,7 @@
 #ifndef FEERTOS_TIMER_H
 #define FEERTOS_TIMER_H
 
+#include <stdint.h>
 #include <stdbool.h>
 #include <stdlib.h>
 
@@ -14,7 +15,7 @@ typedef struct SFeeRTOS_Timer* TFeeRTOS_TimerHandle;
  * aAutoReload: true = periodisch (Timer laeuft nach Callback weiter),
  *              false = einmalig (One-Shot, Timer wird nach Callback geloescht)
  */
-TFeeRTOS_TimerHandle FeeRTOS_CreateTimer(unsigned long aMillis, void (*callback)(void* args), void* args, bool aAutoReload);
+TFeeRTOS_TimerHandle FeeRTOS_CreateTimer(uint32_t aMillis, void (*callback)(void* args), void* args, bool aAutoReload);
 
 void FeeRTOS_DeleteTimer(TFeeRTOS_TimerHandle timer);
